Added password hashing
Password for user is hashed upon creation, the salt used for hashing is also stored in the user row.
The main file accesses the hashed password and the salt with the userID that the user enters to check if the password matches.
Honestly, not sure if this will work immediately